Remote Sensing Segmentation/requirements.txt
torch 
torchvision
segmentation-models-pytorch
albumentations
opencv-python
matplotlib 
seaborn
pandas 
numpy 
scipy
tqdm 
pillow
rasterio  # For remote sensing data